Function shard::provider::dao

pub fn dao(
    db_path: Option<String>
) -> Result<Arc<Mutex<Box<dyn ShareEntryDaoTrait>>>, Box<dyn Error>>

Creates and returns a DAO instance based on the specified database path.

If a path is provided, a Sled database DAO is created; otherwise, an in-memory HashMap DAO is used. This allows flexibility in choosing the underlying storage mechanism.

Arguments

db_path - An optional string slice representing the path to the database.

Returns

Returns a Result<Arc<Mutex<Box<dyn ShareEntryDaoTrait>>>>, encapsulating the DAO in a thread-safe, reference-counted pointer, or an error if the database cannot be initialized.