Trait shard::repository::ShareEntryDaoTrait

pub trait ShareEntryDaoTrait: Send + Sync {
    // Required methods
    fn insert(
        &self,
        key: &str,
        entry: &ShareEntry
    ) -> Result<(), Box<dyn Error>>;
    fn get(&self, key: &str) -> Result<Option<ShareEntry>, Box<dyn Error>>;
    fn get_all(&self) -> Result<Vec<(String, ShareEntry)>, Box<dyn Error>>;
    fn update(
        &self,
        key: &str,
        entry: &ShareEntry
    ) -> Result<(), Box<dyn Error>>;
    fn delete(&self, key: &str) -> Result<(), Box<dyn Error>>;
}

Defines the Data Access Object (DAO) trait for ShareEntry.

This trait specifies the methods for inserting, retrieving, updating, and deleting ShareEntry objects in a data store.

Required Methods

fn insert(&self, key: &str, entry: &ShareEntry) -> Result<(), Box<dyn Error>>

Inserts a ShareEntry into the data store.

Arguments

key - The key associated with the ShareEntry.

entry - The ShareEntry to be inserted.

Returns

A Result indicating the success or failure of the operation.

fn get(&self, key: &str) -> Result<Option<ShareEntry>, Box<dyn Error>>

Retrieves a ShareEntry from the data store by its key.

Arguments

key - The key of the ShareEntry to retrieve.

Returns

A Result containing an Option<ShareEntry>. None if the key does not exist.

fn get_all(&self) -> Result<Vec<(String, ShareEntry)>, Box<dyn Error>>

fn update(&self, key: &str, entry: &ShareEntry) -> Result<(), Box<dyn Error>>

Updates an existing ShareEntry in the data store.

Arguments

key - The key of the ShareEntry to update.

entry - The new ShareEntry data.

Returns

A Result indicating the success or failure of the operation.

fn delete(&self, key: &str) -> Result<(), Box<dyn Error>>

Deletes a ShareEntry from the data store by its key.

Arguments

key - The key of the ShareEntry to delete.

Returns

A Result indicating the success or failure of the operation.