Function shard::sss::generate_refresh_key

pub fn generate_refresh_key(
    threshold: usize,
    secret_length: usize
) -> Result<Vec<Polynomial>, String>

Generates a set of polynomials for refreshing shares in a proactive secret sharing scheme.

Each polynomial generated has a zero constant term, ensuring that when used for refreshing, they do not alter the underlying secret. This function is critical in scenarios where the security of the shares needs to be periodically enhanced.

Arguments

threshold - The minimum number of shares required to reconstruct the secret.

secret_length - The length of the secret in bytes.

Returns

A Result containing a vector of Polynomial objects if successful, or an error message.

Errors

Returns Err if the threshold is less than or equal to 1.

Examples

Generating polynomials for refreshing shares:

let polynomials = generate_refresh_key(3, 5).unwrap();
// `polynomials` now contains 5 polynomials for refreshing shares.