Function shard::sss::refresh_share

pub fn refresh_share(
    share: (&u8, &mut Vec<u8>),
    polynomials: &[Polynomial]
) -> Result<(), String>

Refreshes a single share in a proactive secret sharing scheme.

This function updates the value of a given share by adding new values generated from a set of polynomials. This is useful in scenarios where the security of shares needs to be periodically enhanced without changing the secret.

Arguments

share - A tuple containing the share’s identifier and a mutable reference to the share’s value.

polynomials - A slice of Polynomial objects used to generate new values for refreshing the share.

Returns

A Result<(), String> indicating successful completion or an error message.

Errors

Returns Err if the share is empty.

Returns Err if the length of the share does not match the number of polynomials.

Examples

Refreshing an individual share:

let mut share = (1, vec![5u8, 10]);
let polynomials = generate_refresh_key(3, 2).unwrap();
refresh_share(&mut share, &polynomials).unwrap();
// The share is now updated with new values.